[FEATURE/VG-372] 맵 매니저 및 스테이지 코드 개선 및 레벨 변수 추가

+++ Source/GA6thFinal_Framework/GameScripts/Scripts/Monster/Common/MonsterCommon.cpp
@@ -122,7 +122,7 @@ namespace Monster
 
         return result;
     }
-    SpawnID GetSpawnID(size_t mainLevelIndex, size_t subLevelIndex, size_t battleCount)
+    SpawnID GetSpawnID(int mainLevelIndex, int subLevelIndex, int battleCount)
     {
         std::string spawnIDStr = std::string(SPAWN_ID_HEADER);
         spawnIDStr += std::to_string(mainLevelIndex);
